restructure modules
