+++ src/lib.rs
@@ -1,2 +1,10 @@
-pub mod control_packet;
+//! # mqtt-packet
+//!
+//! A [MQTT v5.0][mqtt] packet parser and generator. This crate is unstable and under active development.
+//!
+//! [mqtt]: https://docs.oasis-open.org/mqtt/mqtt/v5.0/os/mqtt-v5.0-os.html
+//! [rust]: https://www.rust-lang.org/
+
 pub mod data_type;
+pub mod packet_type;
+pub mod property;
